.gemini/antigravity/scratch/CLI-agent-claude-cli-agent-planning-Z804h/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurex"
version = "0.1.0"
description = "AurexAI Python backend - LLM, search, cache"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.setuptools.packages.find]
include = ["aurex*"]
